// apps/website-example/src/app/types.ts
export interface ImageData {
  url: string
  width: number
  height: number
  description?: string
  caption?: string
  source?: string
  focusPoint?: {x: number; y: number}
  format: string

  ogURL: string
  smallTeaserURL: string
  mediumTeaserURL: string
  largeURL: string
}
// ...
export interface Peer {
  id: string
  slug: string
  name: string
  logoURL: string
  isDisabled?: boolean | null
  websiteURL: string
  themeColor: string
  themeFontColor: string
  callToActionText: Node[]
  callToActionURL: string
  callToActionImage: string
  callToActionImageURL: string
}